transport: make _requests_serving param uint32_t
It's not realistic for a shard to have over 4 billion concurrent
requests, so this value can be safely represented in 32 bits.
Also, since the current concurrency limit is represented in uint32_t,
it makes sense for these two to have matching types.

+++ transport/server.hh
@@ -132,7 +132,7 @@ private:
     uint64_t _connects = 0;
     uint64_t _connections = 0;
     uint64_t _requests_served = 0;
-    uint64_t _requests_serving = 0;
+    uint32_t _requests_serving = 0;
     uint64_t _requests_blocked_memory = 0;
     uint64_t _requests_shed = 0;
     auth::service& _auth_service;
